Create folder for scoring functions
